refactor: Update getLiquidationPlans function to use database queries
In order to fetch real-world data, the dummy data return implementation in the getLiquidationPlans function has been replaced with actual database queries. This refactoring introduces usage of 'drizzle-orm' for queries and interactions with the database. The `to` and `from` fields in Compensation interface have also been updated to allow null values.